### 2019-01-05 #1154 @whedon
[PRE REVIEW]: QMRTools: a mathematica toolbox for quantitative MRI analysis
**Submitting author:** @mfroeling (<a href="http://orcid.org/0000-0003-3841-0497">Martijn Froeling</a>)
**Repository:** <a href="https://github.com/mfroeling/QMRITools" target ="_blank">https://github.com/mfroeling/QMRITools</a>
**Version:** v2.0
