+++ paper/paper.md
@@ -44,7 +44,7 @@ The library of functions grows along with the research it is used for. The toolb
 
 The toolbox contains some basic functionality such as DICOM and Nifti import, and 2D,3D and 4D data visualization. 
 
-![Overview](https://github.com/mfroeling/QMRITools/blob/master/Images/OverView.jpg)
+![Overview](OverView.png)
 
 The advanced features comprise data registration (@Klein2010, @Shamonin2013), Noise suppression (@Aja-Fernandez2008, @Veraart2015), Diffusion drift correction (@Vos2014), gradient direction optimization (@Froeling2016), simulation framework (@Froeling2013), EPG based T2 fitting (@Marty2016 @Weigel2015) and IDEAL Dixon reconstruction (@Reeder2005 @Herraez2002).
 
