+++ paper/paper.bib
@@ -1,7 +1,7 @@
 % Encoding: UTF-8
 
 @Article{Froeling2012,
-  author    = {Froeling, Martijn and Nederveen, Aart J. A. J. and Heijtel, Dennis F. R. D. F. R. and Lataster, Arno and Bos, Clemens and Nicolay, Klaas and Maas, Mario and Drost, M. R. Maarten R. and Strijkers, G. J. Gustav J.},
+  author    = {Froeling, Martijn and Nederveen, Aart J. and Heijtel, Dennis F. R. and Lataster, Arno and Bos, Clemens and Nicolay, Klaas and Maas, Mario and Drost, Maarten R. and Strijkers, Gustav J.},
   title     = {{Diffusion-tensor MRI reveals the complex muscle architecture of the human forearm}},
   journal   = {Journal of Magnetic Resonance Imaging},
   year      = {2012},
@@ -83,7 +83,7 @@
 }
 
 @Article{Froeling2016,
-  author   = {Froeling, Martijn and Tax, Chantal M. W. C. M. W. and Vos, Sjoerd B. S. B. and Luijten, Peter R. P. R. and Leemans, Alexander},
+  author   = {Froeling, Martijn and Tax, Chantal M. W. and Vos, Sjoerd B. and Luijten, Peter R. and Leemans, Alexander},
   title    = {{MASSIVE brain dataset: Multiple acquisitions for standardization of structural imaging validation and evaluation}},
   journal  = {Magnetic Resonance in Medicine},
   year     = {2017},
@@ -100,7 +100,7 @@
 }
 
 @Article{Vos2014,
-  author   = {Vos, Sjoerd B. S. B. and Tax, Chantal M. W. C. M. W. and Luijten, Peter R. P. R. and Ourselin, Sebastien and Leemans, Alexander and Froeling, Martijn},
+  author   = {Vos, Sjoerd B. and Tax, Chantal M. W. and Luijten, Peter R. and Ourselin, Sebastien and Leemans, Alexander and Froeling, Martijn},
   title    = {{The importance of correcting for signal drift in diffusion MRI}},
   journal  = {Magnetic Resonance in Medicine},
   year     = {2017},
@@ -163,7 +163,7 @@
 }
 
 @Article{Hooijmans2015,
-  author   = {Hooijmans, M. T. T. and Damon, B. M. M. and Froeling, M. and Versluis, M. J. J. and Burakiewicz, J. and Verschuuren, J. J. G. M. J. G. M. and Niks, E. H. H. and Webb, A. G. G. and Kan, H. E. E.},
+  author   = {Hooijmans, M. T. and Damon, B. M. and Froeling, M. and Versluis, M. J. and Burakiewicz, J. and Verschuuren, J. J. G. M. and Niks, E. H. and Webb, A. G. and Kan, H. E.},
   title    = {{Evaluation of skeletal muscle DTI in patients with duchenne muscular dystrophy}},
   journal  = {NMR in Biomedicine},
   year     = {2015},
@@ -178,7 +178,7 @@
 }
 
 @Article{Froeling2013,
-  author   = {Froeling, Martijn and Nederveen, Aart J. A. J. and Nicolay, Klaas and Strijkers, G. J. Gustav J.},
+  author   = {Froeling, Martijn and Nederveen, Aart J. and Nicolay, Klaas and Strijkers, Gustav J.},
   title    = {{DTI of human skeletal muscle: The effects of diffusion encoding parameters, signal-to-noise ratio and T2 on tensor indices and fiber tracts}},
   journal  = {NMR in Biomedicine},
   year     = {2013},
