### 2019-01-08 #1154 @Kevin-Mattheus-Moerman
@cmadan these also seem suitable reviewers (people I contacted for https://github.com/openjournals/joss-reviews/issues/889) nnadeau, davidnsousa, markus-nilsson, JianCheng
