### 2019-01-09 a4292b4
fix that cut data sometimes not finds middle
