### 2019-01-09 #1154 @JoshKarpel
@cMadan Would be happy to, although I don't know anything about MRI/QMR so I won't be able to address any related technical concerns.
