### 2019-01-18 #1154 @rljacobson
@mfroeling I am definitely qualified to review Mathematica code and general computer vision applications, though I do not have specific expertise with MRI imaging. I do not have any expertise in medicine, but I am quite familiar with the mathematical underpinnings of MRI if that helps. Because I start teaching quite soon, my available time is limited. If you willing to wait ~3 weeks for a completed review, I am willing to review it for you. (The guidelines suggest 2 weeks is more typical.)
