### 2019-01-23 #1154 @rljacobson
@mfroeling Unfortunately, my computer vision expert had to decline. I'll try to be expedient.
