### 2019-01-26 #1154 @cMadan
@rljacobson and others tagged earlier, if you're still interested in reviewing this submission, please let me know. This work is at the intersection of a topic and programming language that don't often overlap, so adequate reviewing will likely include people who cover only part of the project's content--though I also wouldn't want there to be too many reviewers.

I'm going to start the actual review now, but can still add an additional reviewer if someone has time and interest, and would lead to a better project when it gets accepted.
