### 2019-01-29 mfroeling/QMRITools#3 @JoshKarpel
Broken markdown in README
Line 78 of the README has some broken markdown: "The toolbox can be loaded by using <<QMRITools`". I assume it's supposed to be "The toolbox can be loaded by using `<<QMRITools`"

(Part of https://github.com/openjournals/joss-reviews/issues/1204)
