### 2019-01-29 mfroeling/QMRITools#4 @JoshKarpel
Installation instruction for external dependencies unnecessary?
Under the "Install toolbox" heading, the README states that I need to download some external dependencies and place them in a specific folder. When I unpacked QMRITools itself those dependencies were already included, so I didn't need to download them separately.

I assume they either shouldn't be automatically included, or the instructions in the README are now unnecessary :)

(Part of https://github.com/openjournals/joss-reviews/issues/1204)
