### 2019-01-29 mfroeling/QMRITools#3 @JoshKarpel
Ah, I see, the actual Mathematica command has a backtick in it. How annoying.

Looks like you can use doubled backticks to get it to display right: https://meta.stackexchange.com/questions/82718/how-do-i-escape-a-backtick-within-in-line-code-in-markdown

`` ``<<QMRITools` `` ``  = ``<<QMRITools` ``
