### 2019-01-29 mfroeling/QMRITools#5 @JoshKarpel
Better default path for data in demo notebook
The demo notebook uses `SetDirectory["D:\\Werk\\Workspace\\QMRITools\\DemoData"];` - perhaps it could be `SetDirectory[FileNameJoin[{NotebookDirectory[], "DemoData"}]];`, which is a better default for people cloning the package via git, with a comment that says to change it if the demo data is somewhere else?

(Part of https://github.com/openjournals/joss-reviews/issues/1204)
