### 2019-01-29 mfroeling/QMRITools#6 @JoshKarpel
Add automated tests
I haven't the faintest idea how testable this kind of code is, but at the very least having some basic sanity checks ("has it been installed correctly?" type checks) would be nice.

It looks like Mathematica does have a built-in testing framework: https://reference.wolfram.com/language/tutorial/UsingTheTestingFramework.html

Perhaps a "tests" notebook could be supplied along the with the "demo" notebook?

(Part of https://github.com/openjournals/joss-reviews/issues/1204)
