+++ README.md
@@ -75,7 +75,7 @@ The toolbox contains a file ``demo.nb`` and demo data folder ``DemoData``. This 
 
 ## Using the toolbox
 
-The toolbox can be loaded by using <<QMRITools`
+The toolbox can be loaded by using `` <<QMRITools` ``
 
 A list of all QMRITools packages is generated by 
 	
