### 2019-01-29 mfroeling/QMRITools#5 @mfroeling
Indeed that should be the correct path specification.

I have changed it in the `demo.nb` and have also added some explanation in the readme
