### 2019-01-29 fe4e84e
Update demo.nd
and readme for demo
