### 2019-01-29 mfroeling/QMRITools#6 @mfroeling
I have also given this a lot of though on how i can check/test the code but I haven't come up with a good solution.

I am aware of the Mathematica testing framework but is seems a bit cumbersome in this case.

A correct install check i can definitely do. It is easy to check if all function with the correct options I expect from my side (i.e. the ones i have defined) are all available after install.

i will write such a function asap.
