### 2019-01-30 864e9f1
Add install test to repro
