### 2019-01-30 mfroeling/QMRITools#6 @mfroeling
i have added a file and install check test and a test after loading to see if all functions are present.

![test](https://user-images.githubusercontent.com/15890391/51990263-f8df4c00-24a8-11e9-946e-1b70303ec106.PNG)

and have added the tests to the demo.nb

to make it more convenient i have moved all demos and testing files to the DemoAndTest directory to make the download more converient.
