### 2019-01-31 #1204 @JoshKarpel
Calling @grlee77 and @rljacobson to https://github.com/mfroeling/QMRITools/issues/6 when you get a chance - any ideas for automated testing?
