### 2019-01-31 #1204 @mfroeling
@JoshKarpel i will make the paper a bit more explanatory with regard to the points you raised.
