### 2019-02-04 mfroeling/QMRITools#6 @grlee77
I think the demos when combined with the install check above will at least test that things are not completely broken. Users are likely to have more confidence in the package if tests can be extended further to test against various possible corner cases, etc.   From what I understand the Mathematica API tends to have good backwards compatibilty across versions? If that is the case, then hopefully the existing functionality will continue to work without substantial modification as new Mathematica releases are made.

Due to the wide variety of functions in this toolkit, I think adding a comprehensive set of unit tests would be a large amount of work. I am also not familiar with how easy/mature the tools for this in Mathematica are to work with. A more limited approach would be to just add specific unit tests corresponding to bugs as they are encountered to prevent the same problems from occurring again in the future.

I assume that QMITools has been developed during use on real-world data, but if this was human data it may not be possible to make it publicly available. As I'm sure you are aware, there are commercially available MRI phantoms for validating things such as fat fraction (Dixon imaging) or with calibrated T1 or T2 values for relaxometry, but these are expensive and require access to MRI scanner time to obtain the data. Due to the above, as well as the potentially large size of real-world data, synthetic test cases are probably more feasible, but can take quite a bit of time/work to implement.

Another option would be to reuse existing test data or adapt test cases from other appropriately licensed software packages. For example, I know that there are pretty extensive tests related to diffusion MRI implemented as part of the [DIPY](http://nipy.org/dipy/) Python package. Reduced data sets (e.g. [3x3x56](https://github.com/nipy/dipy/blob/cfa1ed1401baf2038bd64ef16702e78750c96aff/dipy/reconst/tests/test_dti.py#L319-L326)) are used in some DTI tests.
