### 2019-02-04 mfroeling/QMRITools#7 @grlee77
add literature references for the methods used
Functions and arguments in the libarry tend to be clearly named, but documentation on the underlying implementations used are lacking.  For basic manipulations/visualization functions this is not an issue as they are pretty self-explanatory, but for a number of other functions it would be helpful to specify a reference for the implementation. This can be as simple as a single paragraph summary and/or a link to a corresponding publication.

**Describe the solution you'd like**

It would be helpful to have the following
- A literature reference (ideally including the paper DOI) corresponding to the method implemented.
- If the implementation differs substantially from what is described, provide a brief summary of the difference and rationale behind it.

**Additional context**

As a concrete example from `DixonTools.nb`. A description provided is:
``DixonReconstruct \[LongDash] Performs Dixon reconstruction of real and imaginary data.``

From the name of that function, I would not know if it is performing basic two-point Dixon, IDEAL fat/water estimation or one of a number of other potential approaches based on graph-cuts, etc.? Inspection of the code reveals that it appears to be IDEAL with support for a multi-peak fat model and involves iterative potentially accounting for off-resonance and T2* decay. A potential user should be able to quickly infer what method is going to be applied without having to read the underlying source code.
