### 2019-02-05 mfroeling/QMRITools#7 @mfroeling
Hi indeed that is lacking but can easily be added.

I like you suggestion to add paragraphs to the guide pages of the documentation with links to the DOI of the manuscripts (since i like my function help to be as clear as possible).

I will only do this for the main processing functions scuh as T2 fitting dixon reconstruction denoising etc. that have a solid basis in literature.

I will let you know when they are updated.
