### 2019-02-05 mfroeling/QMRITools#6 @mfroeling
@grlee @JoshKarpel
Mathematica has excellent backward compatibility. I myself only work in the the latest version and as stated i will only provide support in the latest Mathematica version. In the 8 year i have been working on this I never had any issue upgrading to the latest version.

Indeed all my functions are tested on real life data and there are many qMRI phantoms. however phantom data servers the purpose of test the entire MRI setup, machine, sequence, acquisition etc etc. and not only the code used for processing. Even getting phantom data result the same between center using the same setup and software remains a challenge in most MRI applications.

For most complex applications i have also implemented simulation frameworks. For example DTI/DKI and EPG-T2 fitting, phase unwrapping.
One example of the DTI simulations and the ability of the fitting to provide the true values is given in the demo. In theory i could write some very simple fit test on some simulated data but i feel that is a bit cheating since the code generating the signals is easy to fit while real life data is not. And the robustness of most methods rely not only on noise, but also artifacts motion etc.

In the demos file i provide some of those examples, for phase unwrapping, motion correction, EPG signal simulation and DTI fitting.
If you want i could make them a bit more elaborate and clear.
