### 2019-02-05 mfroeling/QMRITools#6 @grlee77
I think the demo you have is good and tests a substantial portion of the more complicated functions. This kind of integration testing certainly has value and as you correctly point out is a better indicator than unit tests of the performance on realistic data. I also agree with your points on validation via phantoms being complicated. Ultimately the end user should always validate the combination of software + scanner pulse sequence at their site as there are substantial differences among vendors in underlying sequence implementation and the types of pre-processing/corrections applied by the system.

I am not sure offhand roughly what percentage of the code base is covered under this demo, but perhaps you could just fill in some holes for a few things that don't seem to currently be covered (CardiacTools, CoilTools, PhysioTools, etc? unless I missed something).
