### 2019-02-05 mfroeling/QMRITools#7 @grlee77
A couple of other random things I noticed when looking over the code:

1.) PhysioTools: What format(s) of physiology log files is this supposed to process? I think each vendor uses their own format for respiratory, ECG and/or PPU logs.

2.) I saw the term "Overplus" in relation to diffusion gradients. Isn't this a Philips-specific term to refer to gradients played out along a combination of the X/Y/Z axes to allow shorter echo times at a given b-value? I'm not sure users coming from other systems would know what that is referring to.
