### 2019-02-05 mfroeling/QMRITools#6 @mfroeling
@grlee77 will do, i will cover more in the demo
