### 2019-02-06 2d3f5f5
Update Demo
bug fix in installTest
added references to guides
rename DataTransformation
