+++ paper/paper.bib
@@ -1,5 +1,3 @@
-% Encoding: UTF-8
-
 @Article{Froeling2012,
   author    = {Froeling, Martijn and Nederveen, Aart J. and Heijtel, Dennis F. R. and Lataster, Arno and Bos, Clemens and Nicolay, Klaas and Maas, Mario and Drost, Maarten R. and Strijkers, Gustav J.},
   title     = {{Diffusion-tensor MRI reveals the complex muscle architecture of the human forearm}},
