### 2019-02-06 mfroeling/QMRITools#6 @mfroeling
@grlee77 I have added more examples to the data. it now covers around 60% of all functions. However many functions i only use internally in other functions and rarely used by a user. The demos touches around 80% of my code.

The functions covered in the demo are the 99% ones that will be needed in daily use by a normal user. There are many functions that i cannot give demos of since the cover very specific experimental data that is difficult to generalize (mostly Coiltools and PhysioTools).

i will try to cover that in the new paper.md that i am preparing.
