### 2019-02-06 #1204 @grlee77
For reference, the following issues were opened at the source repo:
@JoshKarpel opened an issue related to testing: mfroeling/QMRITools#6
@grlee77 opened an issue related to documentation: mfroeling/QMRITools#7
