### 2019-02-06 mfroeling/QMRITools#6 @grlee77
@JoshKarpel, in the JOSS review you asked "when you get a chance - any ideas for automated testing?"

Did you mean via a continuous integration service? I guess the need for a Mathematica license makes this a little more difficult than for the Python projects I have experience with. @mfoeling could potentially set something up to run on a machine with access to a license using a service like [Jenkins](
https://jenkins.io), but I don't have any direct experience with it or know how much effort it is to set up. This would probably become something that is worth spending time on if the team of contributors grows, but may not be worth it at the moment.

I found the following blog post discussing this related to CI for Matlab which has the same type of commercial license requirements:
http://uwethuemmel.com/continuous-integration-workflow-with-matlab-git-and-jenkins/

Other libraries requiring GPU hardware that is not commonly on the free CI services also use this type of approach (e.g. [CuPy](https://github.com/cupy/cupy)).
