### 2019-02-06 mfroeling/QMRITools#6 @JoshKarpel
@grlee77 A CI service would be nice, but as you mention, needing a Mathematica license makes that pretty non-trivial. I don't think it's necessary for now.

I think I meant "automated" in the sense of "I hit run on this notebook, sit back, and it tells me whether the package works right without me needing to interpret the results by hand". The "demo as integration testing" point noted above is quite reasonable, especially because @mfroeling has increased the demo coverage.
