### 2019-02-06 mfroeling/QMRITools#6 @mfroeling
To make sure that one can check intended behaviour of the three main proccessing algorithms (DTI, EPG-T2, and Dixon) i have now added proper simulations for all three cases.
For that i have extended the T2-EPG simulation framework and added a Dixon (Gradient Echo) simulator.

In the demo i show how people can use these simulations to test the fitting with respect to important parameters such as B1, B0, SNR, fat fraction etc.

Since validation and investigation of limitations of these methods in clinical and multi-center studies are actually ongoing projects im currently working on, I currently cannot provide more code for evaluation of these specific methods since they are WIP, but the will be updated when completed.

However feel free to post any suggestions on how i can further improve testebility of such functions.

![sim-dixon](https://user-images.githubusercontent.com/15890391/52368218-35351e00-2a4e-11e9-8192-5eb961b87fca.PNG)
![sim-t2](https://user-images.githubusercontent.com/15890391/52368219-35351e00-2a4e-11e9-9a2d-2eb89897519e.PNG)
![sim-dti](https://user-images.githubusercontent.com/15890391/52368220-35351e00-2a4e-11e9-85cd-461fa5158604.PNG)
