+++ paper/paper.bib
@@ -1,3 +1,4 @@
+% Encoding: UTF-8
 @Article{Froeling2012,
   author    = {Froeling, Martijn and Nederveen, Aart J. and Heijtel, Dennis F. R. and Lataster, Arno and Bos, Clemens and Nicolay, Klaas and Maas, Mario and Drost, Maarten R. and Strijkers, Gustav J.},
   title     = {{Diffusion-tensor MRI reveals the complex muscle architecture of the human forearm}},
@@ -156,7 +157,7 @@
   issn     = {1522-2594},
   doi      = {10.1002/mrm.26059},
   isbn     = {1522-2594},
-  keywords = {Marchenko–Pastur,Rician,diffusion MRI,noise,principal component analysis},
+  keywords = {Marchenko--Pastur,Rician,diffusion MRI,noise,principal component analysis},
   pmid     = {26599599},
 }
 
@@ -224,6 +225,213 @@
   keywords = {.,Diffusion-weighted imaging (DWI) restoration,Linear minimum mean square error (LMMSE) estimator,Magnetic resonance imaging (MRI),Noise filtering,Rician distribution},
   language = {English},
   pmid     = {18815091},
+}
+
+@Article{Cerqueira2002,
+  author   = {Cerqueira, Manuel D. and Weissman, Neil J. and Dilsizian, Vasken and Jacobs, Alice K. and Kaul, Sanjiv and Laskey, Warren K. and Pennell, Dudley J. and Rumberger, John A. and Ryan, Thomas J. and Verani, Mario S.},
+  title    = {{Standardized myocardial segmentation and nomenclature for tomographic imaging of the heart}},
+  journal  = {Journal of Cardiovascular Magnetic Resonance},
+  year     = {2002},
+  volume   = {4},
+  number   = {2},
+  pages    = {203--210},
+  issn     = {1097-6647},
+  abstract = {... With Heart Disease Receiving Medications for Attention ... A Scientific Statement From the American Heart Association Council on Cardiovascular Disease in the Young Congenital Cardiac Defects Committee and the Council on Cardiovascular Nursing ... $\backslash$n},
+  doi      = {10.1081/JCMR-120003946},
+  isbn     = {8006116083},
+  pdf      = {:C$\backslash$:/Users/mfroelin.LA018015/Desktop/hc0402.102975.pdf:pdf},
+  pmid     = {11815441},
+  url      = {http://www.dekker.com/servlet/product/DOI/10.1081-JCMR-120003946},
+}
+
+@Article{Veraart2016,
+  author        = {Veraart, Jelle and Novikov, Dmitry S. and Christiaens, Daan and Ades-aron, Benjamin and Sijbers, Jan and Fieremans, Els},
+  title         = {{Denoising of diffusion MRI using random matrix theory}},
+  journal       = {NeuroImage},
+  year          = {2016},
+  volume        = {142},
+  pages         = {394--406},
+  month         = nov,
+  issn          = {1053-8119},
+  abstract      = {We introduce and evaluate a post-processing technique for fast denoising of diffusion-weighted MR images. By exploiting the intrinsic redundancy in diffusion MRI using universal properties of the eigenspectrum of random covariance matrices, we remove noise-only principal components, thereby enabling signal-to-noise ratio enhancements. This yields parameter maps of improved quality for visual, quantitative, and statistical interpretation. By studying statistics of residuals, we demonstrate that the technique suppresses local signal fluctuations that solely originate from thermal noise rather than from other sources such as anatomical detail. Furthermore, we achieve improved precision in the estimation of diffusion parameters and fiber orientations in the human brain without compromising the accuracy and spatial resolution.},
+  archiveprefix = {arXiv},
[…]
