### 2019-02-07 #1204 @grlee77
> What kind of people would benefit from using this package over whatever other tools are available?

It appears that engineers and MR physicists as opposed to clinicians/radiologists are the target audience.  Fat/water fraction estimation, diffusion tensor estimation and T2-fitting based on extended phase graphs (EPG) are the three primary applications you have highlighted. All of these are of current interest to researchers in the field of MRI.

There are a number of other widely used software packages for diffusion tensor analysis across a range of programming languages. I am not sure if there are any others implemented in Mathematica, though. I think there is much less current open source software for Dixon fat/water estimation and for EPG-based T2-fitting. Would you agree with this @mfroeling or is there some other unique capability in regards to diffusion imaging (perhaps the parameter estimate vs. SNR simulations?)
