### 2019-02-07 #1204 @grlee77
One other potential issue that should probably be pointed out in the documentation:

Although the software has a permissive BSD 3-clause license, in practice use in commercial applications may be prohibited due to existing patents (I'm pretty sure this is the case for the IDEAL fat/water estimation algorithm?). Each MR system vendor tends to sell their own proprietary closed-source software options for fat/water estimation.
