### 2019-02-07 72e674c
update guide pages with doi numbers
