+++ paper/paper.bib
@@ -48,19 +48,20 @@
 }
 
 @Article{Shamonin2013,
-  author   = {Shamonin, Denis},
-  title    = {{Fast parallel image registration on CPU and GPU for diagnostic classification of Alzheimer's disease}},
-  journal  = {Frontiers in Neuroinformatics},
-  year     = {2013},
-  volume   = {7},
-  number   = {January},
-  pages    = {50},
-  month    = jan,
-  issn     = {1662-5196},
-  doi      = {10.3389/fninf.2013.00050},
-  isbn     = {1662-5196 (Electronic)$\backslash$r1662-5196 (Linking)},
-  keywords = {Acceleration,Alzheimer's disease,Elastix,OpenCL,image registration,parallelization},
-  pmid     = {24474917},
+  author        = {Shamonin, Denis},
+  title         = {{Fast parallel image registration on CPU and GPU for diagnostic classification of Alzheimer's disease}},
+  journal       = {Frontiers in Neuroinformatics},
+  year          = {2013},
+  volume        = {7},
+  number        = {January},
+  pages         = {50},
+  month         = jan,
+  issn          = {1662-5196},
+  __markedentry = {[mfroelin:2]},
+  doi           = {10.3389/fninf.2013.00050},
+  isbn          = {1662-5196 (Electronic)$\backslash$r1662-5196 (Linking)},
+  keywords      = {Acceleration,Alzheimer's disease,Elastix,OpenCL,image registration,parallelization},
+  pmid          = {24474917},
 }
 
 @Article{Herraez2002,
@@ -72,6 +73,7 @@
   number        = {35},
   pages         = {7437},
   issn          = {0003-6935},
+  __markedentry = {[mfroelin:2]},
   archiveprefix = {arXiv},
   arxivid       = {arXiv:1011.1669v3},
   doi           = {10.1364/AO.41.007437},
@@ -82,20 +84,21 @@
 }
 
 @Article{Froeling2016,
-  author   = {Froeling, Martijn and Tax, Chantal M. W. and Vos, Sjoerd B. and Luijten, Peter R. and Leemans, Alexander},
-  title    = {{MASSIVE brain dataset: Multiple acquisitions for standardization of structural imaging validation and evaluation}},
-  journal  = {Magnetic Resonance in Medicine},
-  year     = {2017},
-  volume   = {77},
-  number   = {5},
-  pages    = {1797--1809},
-  month    = may,
-  issn     = {1522-2594},
-  address  = {Milan},
-  doi      = {10.1002/mrm.26259},
-  isbn     = {1522-2594},
-  keywords = {brain dataset,diffusion MRI,evaluation,methods development,modeling,structural MRI},
-  pmid     = {27173617},
+  author        = {Froeling, Martijn and Tax, Chantal M. W. and Vos, Sjoerd B. and Luijten, Peter R. and Leemans, Alexander},
+  title         = {{MASSIVE brain dataset: Multiple acquisitions for standardization of structural imaging validation and evaluation}},
+  journal       = {Magnetic Resonance in Medicine},
+  year          = {2017},
+  volume        = {77},
+  number        = {5},
+  pages         = {1797--1809},
+  month         = may,
+  issn          = {1522-2594},
+  __markedentry = {[mfroelin:2]},
+  address       = {Milan},
+  doi           = {10.1002/mrm.26259},
+  isbn          = {1522-2594},
+  keywords      = {brain dataset,diffusion MRI,evaluation,methods development,modeling,structural MRI},
+  pmid          = {27173617},
 }
 
 @Article{Vos2014,
@@ -114,51 +117,54 @@
   pmid     = {26822700},
 }
 
-@Article{Froeling2015,
[…]
