### 2019-02-07 #1204 @mfroeling
@JoshKarpel I have updated the paper with a small paragraph of each toolbox and added a few lines how the toolbox came about.

@grlee77 @JoshKarpel most indeed my toolbox does what most DTI frameworks can and by now there are many, when i started 10 years ago as a student working on DTI of muscle most of them didnt exist and mathematica was my preferred language. As such i started coding my own scripts and functions and this toolbox is the result of that. I'm sure that there are toolboxes out there that are better and more refined regarding some aspects. For my personal need (mostly muscle and cardiac) there arent any tools that allowed me to perform the tasks i wanted. for dixon there is the matlab based ISMRM dixon toolbox  and for EPG there are non as far as i am aware of.

@grlee77 yes IDEAL is patented, so is Dixon, and so is DTI. Do you think however all methods are openly published and can be used. I'm not sure what the consequences are for people that want to use this in commercial settings and if that is something i should care about. it is the one that want to commertially exploit this who has to take care of that or am i mistaken. would a ohter license be more suited?
