### 2019-02-07 mfroeling/QMRITools#7 @mfroeling
@grlee77 i have added around 20 dois troughout the guides in the documentations, it is currently compiling and will upload it once it is done. for completeness i have also generated a PDF of the function documentation such that is readable outside mathematica.
https://github.com/mfroeling/QMRITools/blob/master/QMRITools/All-Functions.pdf
With the  new paper i tried to make clear what is and what is not in the demo file, since some tools such as the physlogging i hardly ever use my self.

indeed adding what input (1D, 2D, 3D or 4D) for each function and what units would make some more clear. However the full documentation is allmost 100 pages so i have to see if i can find the time (and willingness) to go through all and figure out where such additions are needed.

1) yes every vendor uses their own format for phys logging. Im not even sure if the current fucntion stille works on the latest philips log. I havent used the functionality in a while.
2) indeed overpuls is a term i adopted from the philips platform, and is used in generateGradient(GUI) not sure how else to call or explain this, but the visual representation of the gradients show what it is.
