### 2019-02-08 mfroeling/QMRITools#7 @grlee77
Thanks for adding the DOIs.  Having a PDF is nice for reference outside of Mathematica. I don't know if exporting to HTML is similarly easy, but hosting the docs somewhere that they get indexed by Google, etc., might help potential users find your software.

As far as adding dimensions/units, perhaps just start with a few of the main functions likely to be of interest to end users such as Dixon, T2 fitting and DTI? (I am not saying you HAVE to do this for purposes of the review, it is just a suggestion to improve usability).
