### 2019-02-08 #1204 @grlee77
> @grlee77 yes IDEAL is patented, so is Dixon, and so is DTI. Do you think however all methods are openly published and can be used. I'm not sure what the consequences are for people that want to use this in commercial settings and if that is something i should care about. it is the one that want to commertially exploit this who has to take care of that or am i mistaken. would a ohter license be more suited?

I am not an IP/patent law expert, but I don't think you need to change the license of your code. I was just suggesting that it might be worth having a disclaimer warning users that some of the implemented methods may be covered under patent law. An example of this sort of disclaimer from the OpenCV project can be seen here:
https://github.com/opencv/opencv_contrib/blob/d6895a1b25d46af8f820faefe729a325629d0af1/modules/surface_matching/PATENTS.txt#L12-L29
