### 2019-02-08 mfroeling/QMRITools#7 @mfroeling
@grlee77 will defininatly add this for the main functions. Yes i can generate an html for the documentation using the mathematica workbench, and will see if it is possible to host this online with the repository.
