### 2019-02-09 #1204 @JoshKarpel
@mfroeling re: paper

Love the update - I guess I was wrong about it dramatically increasing the length, but only because you added a lot of good information!

A few nitpicks...
- I think you've got a typo in the second paragraph, `mulsce` -> `muscle`.
- Fig 2 seems unnecessary. The paper doesn't/shouldn't include any information about installation.
