### 2019-02-10 mfroeling/QMRITools#6 @grlee77
Yes. Thanks for expanding the demos @mfroeling. Based on the description based in JOSS's review criteria pasted below, I would say this falls under the OK category.

```
Good: An automated test suite hooked up to an external service such as Travis-CI or similar
OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g. a sample input file to assert behaviour)
Bad (not acceptable): No way for you the reviewer to objectively assess whether the software works
```
