### 2019-02-10 mfroeling/QMRITools#7 @grlee77
I see the DOIs were added to various files in your recent commit to master, but I do not see them in the PDF you linked above. Was that built prior to adding them or am I missing something?
