### 2019-02-11 mfroeling/QMRITools#6 @rljacobson
I am happy with this as well.

I just want to add a few references to some useful information about using the Mathematica testing framework:

- [Best practices for using the Testing Framework](https://mathematica.stackexchange.com/questions/170947/best-practices-for-using-the-testing-framework) on SE.
- A convenient [Mathematica test runner](https://github.com/lynchs61/Mathematica-Test-Runner) that is able to produce different output formats compatible with other popular frameworks in addition to JSON and raw WL output.
