### 2019-02-11 d535a12
Some minor editorial suggestions.
