### 2019-02-11 mfroeling/QMRITools#7 @mfroeling
@grlee, indeed i have updated the guide pages which are not included in the PDF. im working on the function files this week including the units where apropirate and also including the DOI on a function level.

once done i well recompile the help and regenerate the PDF which then should include the DOIs.
