### 2019-02-11 mfroeling/QMRITools#6 @mfroeling
@rljacobson i will look into this. It look as an easy thing to implement.
Based on the my initial experience with the testing framework during this review i will definately see how to integrate this more.
I keep the issue open for now and see if i can enhance based on the articles you referenced.
