### 2019-02-11 #1204 @mfroeling
@JoshKarpel indeed after seeing the compiled PDF it needs an other good once over. i was a bit messy, and indeed the figure 2 can go. Based on the comment i got from @rljacobson i will give the paper an other update.

@grlee77 i will add such a disclaimer.
