### 2019-02-22 1cbe4dc
Change of elastix parameter files to include all PCA metrics
