+++ paper/paper.bib
@@ -18,33 +18,35 @@
 }
 
 @Article{Weigel2015,
-  author    = {Weigel, Matthias},
-  title     = {{Extended phase graphs: Dephasing, RF pulses, and echoes - pure and simple}},
-  journal   = {Journal of Magnetic Resonance Imaging},
-  year      = {2015},
-  volume    = {41},
-  number    = {2},
-  pages     = {266--295},
-  month     = feb,
-  issn      = {1053-1807},
-  doi       = {10.1002/jmri.24619},
-  keywords  = {Fourier space,configuration states,dephasing,extended phase graph,partitioning,phase graph},
-  publisher = {Wiley-Blackwell},
+  author        = {Weigel, Matthias},
+  title         = {{Extended phase graphs: Dephasing, RF pulses, and echoes - pure and simple}},
+  journal       = {Journal of Magnetic Resonance Imaging},
+  year          = {2015},
+  volume        = {41},
+  number        = {2},
+  pages         = {266--295},
+  month         = feb,
+  issn          = {1053-1807},
+  __markedentry = {[mfroelin:2]},
+  doi           = {10.1002/jmri.24619},
+  keywords      = {Fourier space,configuration states,dephasing,extended phase graph,partitioning,phase graph},
+  publisher     = {Wiley-Blackwell},
 }
 
 @Article{Marty2016,
-  author   = {Marty, Benjamin and Baudin, Pierre Yves and Reyngoudt, Harmen and Azzabou, Noura and Araujo, Ericky C. A. and Carlier, Pierre G. and de Sousa, Paulo L.},
-  title    = {{Simultaneous muscle water T2and fat fraction mapping using transverse relaxometry with stimulated echo compensation}},
-  journal  = {NMR in Biomedicine},
-  year     = {2016},
-  volume   = {29},
-  number   = {4},
-  pages    = {431--443},
-  issn     = {1099-1492},
-  doi      = {10.1002/nbm.3459},
-  isbn     = {1099-1492},
-  keywords = {Extended phase graph,Fat fraction,Fat infiltration,MSME sequence,Skeletal muscle,Water T2},
-  pmid     = {26814454},
+  author        = {Marty, Benjamin and Baudin, Pierre Yves and Reyngoudt, Harmen and Azzabou, Noura and Araujo, Ericky C. A. and Carlier, Pierre G. and de Sousa, Paulo L.},
+  title         = {{Simultaneous muscle water T2and fat fraction mapping using transverse relaxometry with stimulated echo compensation}},
+  journal       = {NMR in Biomedicine},
+  year          = {2016},
+  volume        = {29},
+  number        = {4},
+  pages         = {431--443},
+  issn          = {1099-1492},
+  __markedentry = {[mfroelin:2]},
+  doi           = {10.1002/nbm.3459},
+  isbn          = {1099-1492},
+  keywords      = {Extended phase graph,Fat fraction,Fat infiltration,MSME sequence,Skeletal muscle,Water T2},
+  pmid          = {26814454},
 }
 
 @Article{Shamonin2013,
@@ -102,22 +104,23 @@
 }
 
 @Article{Vos2014,
-  author   = {Vos, Sjoerd B. and Tax, Chantal M. W. and Luijten, Peter R. and Ourselin, Sebastien and Leemans, Alexander and Froeling, Martijn},
-  title    = {{The importance of correcting for signal drift in diffusion MRI}},
-  journal  = {Magnetic Resonance in Medicine},
-  year     = {2017},
-  volume   = {77},
-  number   = {1},
-  pages    = {285--299},
-  month    = jan,
-  issn     = {1522-2594},
-  doi      = {10.1002/mrm.26124},
[…]
