### 2019-02-25 a1329fe
Add doi numbers to help files
changed method for determining slice outliers for cardiac data

untrac demo data and build folders
