+++ build/QMRITools/PacletInfo.m
@@ -1,19 +1,0 @@
-(* Paclet Info File *)
-
-(* created 2018/12/24*)
-
-Paclet[
-    Name -> "QMRITools",
-    Version -> "2.0.1",
-    MathematicaVersion -> "11.0+",
-    Description -> "Toolbox for Qunatitative MRI.",
-    Creator -> "Martijn Froeling <m.froeling@gmail.com>",
-    Thumbnail -> "zthumb.gif",
-    Extensions -> 
-        {
-            {"Kernel", Root -> "Kernel", Context -> "QMRITools`"}, 
-            {"Documentation", Language -> "English", MainPage -> "Guides/QMRITools"}
-        }
-]
-
-
