### 2019-02-26 0429ac5
Update DOI documentation.
update readme and paper

{AnisoFilterTensor, BullseyePlot, CreateDiffData, DeNoise, \
DixonReconstruct, DriftCorrect, EPGSignal, EPGT2Fit, \
GenerateGradients, GenerateGradientsGUI, HelixAngleCalc, PCADeNoise, \
RegisterData, SimHamiltonian, Tensor, TensorCalc, Unwrap}
