+++ paper/paper.md
@@ -23,11 +23,9 @@ bibliography: paper.bib
 
 The toolbox was developed mostly in the context of quantitative muscle, nerve and cardiac magnetic resonance imaging. The library of functions grows along with the research it is used for and started as a toolbox to analyze DWI data of muscle (@Froeling2012). Although there exist many different software packages and code repositories for much of the functionality in this toolbox, I was in need of one that did all. Furthermore, most diffusion packages are optimized for brain analysis and provide limited flexibility. QMRITools works alongside other software packages (e.g. [vIST/e](http://bmia.bmt.tue.nl/software/viste/), and [ITKSnap](http://www.itksnap.org/pmwiki/pmwiki.php)) and for some functionality it calls external executables (e.g. [dcm2nii](https://www.nitrc.org/projects/dcm2nii/) and [Elastix](http://elastix.isi.uu.nl/)). The toolbox has been used in various studies (e.g. @Froeling2012, @Hooijmans2015, @Froeling2015).
 
-![Overview](overview.png)
+![An overview of some of the functionality in the QMRITools Mathematica add-on.](overview.png)
 
 The toolbox includes some [demo data](https://github.com/mfroeling/QMRITools/releases/download/2.0/DemoAndTest.zip) which is used in the demo file ``demo.nb``. In this notebook most of the functionality of the toolbox is demonstrated. A full list of functions and packages can be found in the [file](https://github.com/mfroeling/QMRITools/blob/master/QMRITools/All-Functions.nb) ``All-Functions.nb`` (also availible as [pdf](https://github.com/mfroeling/QMRITools/releases/download/2.0/All-Functions.pdf)). For all functions and toolboxes, help files and guides are available and the documentation is build such that is incorporated in the Wolfram documentation.  
-
-![Install](addons.png)
 
 The toolbox contains some basic functionality such as DICOM and Nifti import, and 2D, 3D and 4D data visualization. The advanced features comprise data registration (@Klein2010, @Shamonin2013), noise suppression (@Aja-Fernandez2008, @Veraart2015), diffusion drift correction (@Vos2014), gradient direction optimization (@Froeling2016), simulation framework (@Froeling2013), EPG based T2 fitting (@Marty2016 @Weigel2015) and IDEAL Dixon reconstruction (@Reeder2005 @Herraez2002). The current functional toolboxes with a short description are listed below.
 
@@ -47,7 +45,7 @@ Most of the functionality is demonstrated in the `demo.nb`.
 An IDEAL based Dixon reconstruction algorithm (@Reeder2005, @Yu2008). The method provides multi-peak fitting B0 field and T2* correction. The toolbox also provides a function for unwrapping phase data in 2D and 3D based on a best path method (@Abdul-Rahman2007, @Herraez2002). It also contains a function that allows simulating gradient echo Dixon data.
 Most of the functionality is demonstrated in the `demo.nb`.
 
-![dixon](dixon.png)
+![IDEAL based Dixon reconstruction: fitted fat fractions as a function of the imposed fat fraction, SNR and B0 field offset.](dixon.png)
 
 ## ElastixTools
 A wrapper that calls the Elastix registration framework (@Klein2010, @Shamonin2013). The toolbox determines what registration or transformations need to be performed, exports the related data to a temp folder and calls an automatically generated command line script that performs the registration. After registration is completed the data is again loaded into Mathematica.
@@ -61,7 +59,7 @@ Most of the functionality is demonstrated in the `demo.nb`.
[…]
