### 2019-02-26 mfroeling/QMRITools#7 @mfroeling
@grlee the DOI numbers are now also included in the code as well as in the documentation and PDF file.
I still have to do a recompile of the reference page database and will see if the HTML build works.

adding units to the help files is next on my list.
