### 2019-02-26 mfroeling/QMRITools#6 @mfroeling
@rljacobson looked into the references you provide, very usefull!! For now i am keeping what i have but this functionality i will definitely keep in mind for upcoming changes.

@rljacobson @grlee77 @JoshKarpel since you are all happy with this i will close the issue. Feel free to open a new one if needed.
