### 2019-02-26 mfroeling/QMRITools#10 @rljacobson
"File not found" on $TMP/QMRIToolsReg/result-3D.nii.gz.
**Describe the bug**
`RegisterData`, `RegisterDataTransformSplit`, `RegisterDataSplit` fail with an import error. It cannot find `result-3D.nii.gz` in the temporary directory.

**To Reproduce**
Steps to reproduce the behavior:
1. Open the `demo.nb` file and initialize.
2. Scroll to the Registration using Elastix section and import the data.
3. Continue evaluating the code cells under Affine Transformation until you get to `RegisterData` which will fail with an error.

 - OS: macOS
 - Mathematica version 11.3

Other functions in the `Register*` family fail with the same error. Of course, any demonstration code depending on the output of these functions also fail.
