### 2019-02-26 mfroeling/QMRITools#11 @rljacobson
Text on buttons obscured in the GradientsGUI
**Describe the bug**
See attached screenshot. The labels on the buttons cannot be read.

**To Reproduce**
Steps to reproduce the behavior:
1. Scroll to the section DTI and IVIM
2. Evaluate all cells in the Generating diffusion gradients section.

**Screenshots**

**Additional context**
I have a 27inch high dpi iMac display.
