### 2019-02-26 mfroeling/QMRITools#12 @rljacobson
Minor Issues and Suggestions
This is a list of minor issues and suggestions.

- [x] `CompilebleFunctions` The documentation says it "generates a list of all compilable functions." It generates a table (a list of lists wrapped in `TableForm`) of compilable functions, not a `List`.  The function `` Compile`CompilerFunctions `` generates a list of compilable functions.
- [x] `SegmentMask[mask, n]` It is not clear what `SegmentMask` is doing. The docs say, "divides a mask in n equal segments along the slice direction." What does "equal segments" mean? Equal in area? Equal in height? The plot is of a color quantized image, but the legend is continuous and in unknown units.
- [x] For the ME-SE  (TSE) T2 mapping slice profiles, in `SimulateSliceEPG`, the axes are not labeled.

The first one is picking nits. The second one might be my ignorance of the field of application.
