### 2019-02-26 #1204 @rljacobson
@cMadan @mfroeling I am nearly done my review. I have opened the following issues:

1. Issue [#10](https://github.com/mfroeling/QMRITools/issues/10) is a bug preventing the registration functionality from working. I also couldn't test the code depending on the output of those functions in `demo.nb`.
2. Issue [#11](https://github.com/mfroeling/QMRITools/issues/11) is a minor bug with an easy fix.
3. Issue [#12](https://github.com/mfroeling/QMRITools/issues/12) contains either suggestions or very minor issues which are not barriers to acceptance, in my view.
