### 2019-02-27 mfroeling/QMRITools#10 @mfroeling
Indeed the mac compatibility has some issues at the moment. I was aware of this and working on a fix with some collaborators. Currently i don't have any mac system available for testing.

I am working on a full revision of the registrtation framework such that it will be compatible with both linux and Max. currently the software generates windows command line functions and bat files. these have to be converted to bash.

I will let you know when this issue is fixed but i might take a while.
