### 2019-02-27 mfroeling/QMRITools#11 @mfroeling
I have fixed the font size of the button to 10pt (the size used on my windows system) and shortened the text in the buttons hopefully this will fix the issue. will commit asap
I cannot test this since i dont have a mac.
