### 2019-02-27 mfroeling/QMRITools#12 @mfroeling
thanks for the valuable comments.

* I have updated the help description of CompilebleFunctions to
"generates a formatted table of all compilable functions generated by Compile`CompilerFunctions."

* I have updated the help description of SegmentMask to
"divides a mask in n segments along the slice direction, n must be an integer. The mask is divided in n equal parts where each parts has the same number of slices."

* I have added labels to the SimulateSliceEPG plot
![epg](https://user-images.githubusercontent.com/15890391/53478807-f6581e00-3a77-11e9-9b84-7c5a5c53ca57.PNG)
