### 2019-02-27 57f3bc2
Fix issues for paper
Fixed font size in GenerategradientsGUI
added axes labels to Generate gradients
changes help of CompilebleFunctions
changed help of SegmentMask

add disclaimer about patents
