+++ README.md
@@ -166,4 +166,17 @@ A summary of the core functionality is listed below.
 ## License
 https://opensource.org/licenses/BSD-3-Clause
 
+Note that restrictions imposed by these patents (and possibly others)
+exist independently of and may be in conflict with the freedoms granted
+in BSD-3-Clause license, which refers to copyright of the program, not patents
+for any methods that it implements. Both copyright and patent law must
+be obeyed to legally use and redistribute this program and it is not the
+purpose of this license to induce you to infringe any patents or other
+property right claims or to contest validity of any such claims.  If you
+redistribute or use the program, then this license merely protects you
+from committing copyright infringement.  It does not protect you from
+committing patent infringement.  So, before you do anything with this
+program, make sure that you have permission to do so not merely in terms
+of copyright, but also in terms of patent law.
+
 Some code in the NiiTools packages was based on https://github.com/tomdelahaije/nifti-converter
