### 2019-02-28 mfroeling/QMRITools#10 @mfroeling
Unix and Mac compatibility should be OK now, the code is changed still need tot test it. will know more next week.
