### 2019-03-01 mfroeling/QMRITools#7 @mfroeling
@grlee i have created a new release that also includ a html version of the documentation, I have also added all DOIs and units of the parameters where apropriate.

https://github.com/mfroeling/QMRITools/releases/tag/2.1
