### 2019-03-03 #1204 @cMadan
@mfroeling, I understand that there are still a few things left to finalise? (E.g., https://github.com/mfroeling/QMRITools/issues/10)

@grlee77, to the degree that you can evaluate the submission, are you happy to sign-off on it?

@JoshKarpel, is it fair to say you're good with the submission now?

@rljacobson, I think there are still a few minor issues to resolve, correct?
