### 2019-03-04 #1204 @grlee77
Yes, I am satisfied that the author has made improvements to the documentation and also expanded the demo functions for users to more easily test the functionality and determine the capabilities of the package. I just updated a couple of the checklist items to reflect this.

I did not check anything under the "functionality" section due to lack of a Mathematica license to actually try any of the functions in practice.

Is it generally the convention that reviewers should also write up a final review/summary in this issue as well? (This is my first review for JOSS so I am still learning the conventions here).
