### 2019-03-04 #1204 @mfroeling
@cMadan Yes, the open issue of mac compatilitbily i am trying to fix this week. As mentioned in the review it is difficult for me to provide max support, but hopefully this week i am able to get all the needed dependencies compiled and test it on some mac machines.
