### 2019-03-05 2a8b34e
windows and mac path bug fixes
