+++ PacletInfo.m
@@ -1,12 +1,12 @@
 (* Paclet Info File *)
 
-(* created 2019/02/26*)
+(* created 2019/03/01*)
 
 Paclet[
     Name -> "QMRITools",
     Version -> "2.1.0",
     MathematicaVersion -> "11.0+",
-    Description -> "Toolbox for Qunatitative MRI.",
+    Description -> "Toolbox for Quantitative MRI.",
     Creator -> "Martijn Froeling <m.froeling@gmail.com>",
     Thumbnail -> "zthumb.gif",
     Extensions -> 
