### 2019-03-05 #1204 @cMadan
@grlee77, no problem! There isn't a convention of writing a final summary, particularly because the process tends to vary based on the how the authors and reviewers discuss the submission (how much in the review thread vs. in issues in the project's repo), how quickly discussions go back-and-forth, and other factors. What you've done here is fine, but when it's otherwise been quiet, at the end it's helpful to let me know that you're all set or if there are any issues that seem to be particularly problematic.

@JoshKarpel, perfect, thank you!

@mfroeling, that sounds good--thanks for the update!
