### 2019-03-13 mfroeling/QMRITools#11 @mfroeling
@rljacobson the 2.1 release should have the txt issue fixed. Could you please verify?
