### 2019-03-14 3d97cdb
update of readme and mac bugfix elastix
