### 2019-03-14 mfroeling/QMRITools#12 @mfroeling
@rljacobson can i close this issue?
