### 2019-03-14 #1204 @mfroeling
@cMadan @rljacobson the mac compatibility should be fixed as well as all other open issues.
