+++ docs/about.md
@@ -1,0 +1,5 @@
+---
+theme: jekyll-theme-slate
+---
+
+this is a test
