+++ docs/_config.yml
@@ -1,0 +1,3 @@
+theme: jekyll-theme-slate
+title: Martijn Froeling
+description: List of projects and processing tools related to quantitative MRI of muscle, heart and nerves
