+++ docs/_config.yml
@@ -1,3 +1,9 @@
 theme: jekyll-theme-slate
-title: Martijn Froeling
-description: List of projects and processing tools related to quantitative MRI of muscle, heart and nerves
+
+title:       QMRITools
+name:        QMRITools
+tagline:     QMRITools for Mathematica
+url:         http://mfroeling.github.io/QMRITools
+baseurl:     /QMRITools/
+
+description: Processing tools related to quantitative MRI of muscle, heart and nerves.
