+++ docs/index.md
@@ -1,4 +1,28 @@
-## Welcome to my github page
+## Welcome to QRMITools
 
-Here i will add a list of projects and toolboxes that i use for quantitative MRI imaging of muscle, nerve and heart.
+``QMRITools`` is written in Mathematica using Wolfram Workbench and Eclipse and contains a collection of tools and functions for processing quantitative MRI data. The toolbox does not provide a GUI and its primary goal is to allow for fast and batch data processing, and facilitate development and prototyping of new functions. The core of the toolbox contains various functions for data manipulation and restructuring.
 
+The toolbox was developed mostly in the context of quantitative muscle, nerve and cardiac magnetic resonance imaging. The library of functions grows along with the research it is used for and started as a toolbox to analyze DWI data of muscle
+
+***
+
+##Content
+
+* [Information](#information)
+* [link](\documentation.html){:target="_blank"}
+* [Demonstrations](#demonstrations)
+* [Functionality](#functionality)
+
+***
+
+##Information
+
+***
+
+##Demonstrations
+
+***
+
+##Functionality
+
+***
