+++ docs/index.md
@@ -1,28 +1,124 @@
 ## Welcome to QRMITools
 
-``QMRITools`` is written in Mathematica using Wolfram Workbench and Eclipse and contains a collection of tools and functions for processing quantitative MRI data. The toolbox does not provide a GUI and its primary goal is to allow for fast and batch data processing, and facilitate development and prototyping of new functions. The core of the toolbox contains various functions for data manipulation and restructuring.
+`QMRITools` is written in Mathematica using Wolfram Workbench and Eclipse and contains a collection of tools and functions for processing quantitative MRI data. The toolbox does not provide a GUI and its primary goal is to allow for fast and batch data processing, and facilitate development and prototyping of new functions. The core of the toolbox contains various functions for data manipulation and restructuring.
 
-The toolbox was developed mostly in the context of quantitative muscle, nerve and cardiac magnetic resonance imaging. The library of functions grows along with the research it is used for and started as a toolbox to analyze DWI data of muscle
+The toolbox was developed mostly in the context of quantitative muscle, nerve and cardiac magnetic resonance imaging. The library of functions grows along with the research it is used for and started as a toolbox to analyze DWI data of muscle.
+
+The toolbox is developed for the [Wolfram language](https://www.wolfram.com/language/) and maintained using [Wolfram workbench](https://www.wolfram.com/workbench/) for [eclipse](https://www.eclipse.org/) and runs in the latest version of [Wolfram Mathematica].
+
+![eclipse](\QMRITools\images\eclipse.png) ![wolfram language](\QMRITools\images\wolfram language.png) ![Wolfram Mathematica](\QMRITools\images\wolfram mathematatica.png) ![wolfram workbench](\QMRITools\images\wolfram workbench)
+
+
 
 ***
 
 ## Content
-
-* [Information](#information)
+* [Latest Release](#latest-release)
 * [Documentation](\QMRITools\htmldoc\guide\QMRITools.html){:target="_blank"}
 * [Demonstrations](#demonstrations)
 * [Functionality](#functionality)
 
 ***
 
-## Information
+## Latest Release
+
+The latesest release can be found [here](https://github.com/mfroeling/QMRITools/releases){:target="_blank"}. 
+
+Install the toolbox in the Mathematica UserBaseDirectory > Applications.
+
+	`FileNameJoin[{$UserBaseDirectory, "Applications"}]`
+
+Some functions of QMRITools call on external executables and software.
+These executables need to be present in "QMRITools\Applications" and are included in the release.
+If for any reason you want to use other (older/newer) versions you can replace them but functionality is not guaranteed.
+For the latest version of these tools and their user license please visit their website.
+
+* [dcm2niix](https://github.com/rordenlab/dcm2niix/)
+	* dcm2niix.exe
+* [Elastix](http://elastix.isi.uu.nl/)
+	* elastix.exe
+	* transformix.exe
+
+All functionality is tested under Windows 10 with the latest Mathematica version.
+The Mathematica code is cross platform compatible with the exception of the external tools which are compiled for each OS.
+The toolbox provides compiled versions for each OS but their functionality is not guaranteed.
+The Elastix version used is 4.9 with OpenCL support. Additionally Elastix needs to be compiles with the PCA metrics, all DTI related parameters and all affine related parameters.
+
+Although cross platform compatibility is provided I have only limited options for testing so if any issues arise please let me know.  
 
 ***
 
 ## Demonstrations
 
[…]
