+++ README.md
@@ -7,7 +7,10 @@
  
 ***
 
-[![wolfram language](https://github.com/mfroeling/QMRITools/blob/master/Images/wolfram language.png)](https://www.wolfram.com/language/)   [![wolfram workbench](https://github.com/mfroeling/QMRITools/blob/master/Images/wolfram workbench.jpg)](https://www.wolfram.com/workbench/)   [![eclipse](https://github.com/mfroeling/QMRITools/blob/master/Images/clipse.png)](https://www.eclipse.org/)   [![Wolfram Mathematica](https://github.com/mfroeling/QMRITools/blob/master/Images/wolfram mathematica.png)](http://www.wolfram.com/mathematica/)
+[![wolfram language](https://github.com/mfroeling/QMRITools/blob/master/Images/wolfram language.png)](https://www.wolfram.com/language/)   
+[![wolfram workbench](https://github.com/mfroeling/QMRITools/blob/master/Images/wolfram workbench.jpg)](https://www.wolfram.com/workbench/)   
+[![eclipse](https://github.com/mfroeling/QMRITools/blob/master/Images/eclipse.png)](https://www.eclipse.org/)   
+[![Wolfram Mathematica](https://github.com/mfroeling/QMRITools/blob/master/Images/wolfram mathematica.png)](http://www.wolfram.com/mathematica/)
 
 ## Contents
 
