### 2019-03-22 #1204 @mfroeling
FYI all help is now a nice website.

https://mfroeling.github.io/QMRITools
https://mfroeling.github.io/QMRITools/htmldoc/guide/QMRITools.html
