+++ docs/index-cit.md
@@ -180,9 +180,9 @@ A collection of tools to analyze cardiac data. The main features are
 cardiac shape analysis which allows defining the hard in a local
 myocardial coordinate system which allows quantifying and analyzing
 data. When the cardiac geometry is known there are functions to analyze
-qMRI parameters in the AH17 model (Cerqueira et al. 2002) or perform
-transmural sampling of qMRI parameters. Most of the functionality is
-demonstrated in the `demo.nb`.
+qMRI parameters in the AH17 model (Cerqueira et al.
+[2002](#ref-Cerqueira2002)) or perform transmural sampling of qMRI
+parameters. Most of the functionality is demonstrated in the `demo.nb`.
 
 ### CoilTools
 
@@ -196,23 +196,26 @@ This toolbox is not demonstrated in the `demo.nb`.
 ### DenoiseTools
 
 The toobox provides two algorithms that allow denoising of DWI data. The
-first is based on and LMMSE framework (Aja-Fernandez et al. 2008) and
-the second is based on a random matrix theory and Principal component
-analysis framework (Veraart, Fieremans, and Novikov 2016; Veraart et al.
-2016). Furthermore, it provides an anisotropic filter for denoising the
-estimated diffusion tensor which provides more reliable fiber
-orientation analysis (Lee, Chung, and Alexander 2006). Most of the
-functionality is demonstrated in the `demo.nb`.
+first is based on and LMMSE framework (Aja-Fernandez et al.
+[2008](#ref-Aja-Fernandez2008)) and the second is based on a random
+matrix theory and Principal component analysis framework (Veraart,
+Fieremans, and Novikov [2016](#ref-Veraart2015); Veraart et al.
+[2016](#ref-Veraart2016)). Furthermore, it provides an anisotropic
+filter for denoising the estimated diffusion tensor which provides more
+reliable fiber orientation analysis (Lee, Chung, and Alexander
+[2006](#ref-JeeEunLee)). Most of the functionality is demonstrated in
+the `demo.nb`.
 
 ### DixonTools
 
-An IDEAL based Dixon reconstruction algorithm (Reeder et al. 2005; Yu et
-al. 2008). The method provides multi-peak fitting B0 field and T2\*
-correction. The toolbox also provides a function for unwrapping phase
-data in 2D and 3D based on a best path method (Abdul-Rahman et al. 2007;
-Herraez et al. 2002). It also contains a function that allows simulating
-gradient echo Dixon data. Most of the functionality is demonstrated in
-the `demo.nb`.
+An IDEAL based Dixon reconstruction algorithm (Reeder et al.
+[2005](#ref-Reeder2005); Yu et al. [2008](#ref-Yu2008)). The method
+provides multi-peak fitting B0 field and T2\* correction. The toolbox
+also provides a function for unwrapping phase data in 2D and 3D based on
+a best path method (Abdul-Rahman et al. [2007](#ref-Abdul-Rahman2007);
+Herraez et al. [2002](#ref-Herraez2002)). It also contains a function
+that allows simulating gradient echo Dixon data. Most of the
+functionality is demonstrated in the `demo.nb`.
 
 ![IDEAL based Dixon reconstruction: fitted fat fractions as a function
 of the imposed fat fraction, SNR and B0 field
@@ -221,12 +224,12 @@ offset.](\QMRITools\images\dixon.png)
 ### ElastixTools
 
 A wrapper that calls the Elastix registration framework (Klein et al.
-2010; Shamonin 2013). The toolbox determines what registration or
-transformations need to be performed, exports the related data to a temp
-folder and calls an automatically generated command line script that
-performs the registration. After registration is completed the data is
[…]
