+++ docs/index.md
@@ -59,11 +59,11 @@ Applications.
 `FileNameJoin[{$UserBaseDirectory, "Applications"}]`
 
 Some functions of QMRITools call on external executables and software.
-These executables need to be present in “QMRITools” and are included in
-the release. If for any reason you want to use other (older/newer)
-versions you can replace them but functionality is not guaranteed. For
-the latest version of these tools and their user license please visit
-their website.
+These executables need to be present in &ldquo;QMRITools&rdquo; and are
+included in the release. If for any reason you want to use other
+(older/newer) versions you can replace them but functionality is not
+guaranteed. For the latest version of these tools and their user license
+please visit their website.
 
 -   [dcm2niix](https://github.com/rordenlab/dcm2niix/)
     -   dcm2niix.exe
@@ -294,8 +294,9 @@ supported. This toolbox is not demonstrated in the `demo.nb`.
 ### PlottingTools
 
 A variety of functions for visualization of various data types. The main
-functions are ‘PlotData’ and ‘PlotData3D’ which allow viewing 2D, 3D and
-4D data. Most of the functionality is demonstrated in the `demo.nb`.
+functions are &lsquo;PlotData&rsquo; and &lsquo;PlotData3D&rsquo; which
+allow viewing 2D, 3D and 4D data. Most of the functionality is
+demonstrated in the `demo.nb`.
 
 ### ProcessingTools
 
@@ -330,9 +331,9 @@ is demonstrated in the `demo.nb`.
 
 The original toolbox where the project started. The main functions in
 this toolbox are to fit and evaluate the diffusion tensor model. Various
-fitting methods are implemented (e.g. LLS, NLS, WLLS, and iWLLS). The
-default method is an iterative weighted linear least squares approach
-(Veraart et al. 2013). The tensor fitting also includes outlier
+fitting methods are implemented (e.g.&nbsp;LLS, NLS, WLLS, and iWLLS).
+The default method is an iterative weighted linear least squares
+approach (Veraart et al. 2013). The tensor fitting also includes outlier
 detections using REKINDLE (Tax et al. 2015) and data preparation
 includes drift correction (Vos et al. 2017). Most of the functionality
 is demonstrated in the `demo.nb`.
@@ -350,126 +351,135 @@ of the functionality is demonstrated in the `demo.nb`.
 ## References
 
 Abdul-Rahman, Hussein S., Munther A. Gdeisat, David R. Burton, Michael
-J. Lalor, Francis Lilley, and Christopher J. Moore. 2007. “Fast and
-robust three-dimensional best path phase unwrapping algorithm.” *Applied
-Optics* 46 (26): 6623. <https://doi.org/10.1364/AO.46.006623>.
+J. Lalor, Francis Lilley, and Christopher J. Moore. 2007. &ldquo;Fast
+and robust three-dimensional best path phase unwrapping
+algorithm.&rdquo; *Applied Optics* 46 (26): 6623.
+<https://doi.org/10.1364/AO.46.006623>.
 
 Aja-Fernandez, Santiago, Marc Niethammer, Marek Kubicki, Martha E.
[…]
