+++ docs/index.md
@@ -383,8 +383,8 @@ indices and fiber tracts.&rdquo; *NMR in Biomedicine* 26 (11):
 1339&ndash;52. <https://doi.org/10.1002/nbm.2959>.
 
 Froeling, Martijn, Jos Oudeman, G. J. Gustav J. Strijkers, Mario Maas,
-M. R. Maarten R. Drost, Klaas Nicolay, and Aart J. A. J. Nederveen.
-2015. &ldquo;Muscle Changes Detected with Diffusion-Tensor Imaging after
+M. R. Maarten R. Drost, Klaas Nicolay, and Aart J. A. J. Nederveen. 2015. 
+&ldquo;Muscle Changes Detected with Diffusion-Tensor Imaging after
 Long-Distance Running.&rdquo; *Radiology* 274 (2): 548&ndash;62.
 <https://doi.org/10.1148/radiol.14140702>.
 
@@ -423,8 +423,8 @@ Ericky C. A. Araujo, Pierre G. Carlier, and Paulo L. de Sousa. 2016.
 transverse relaxometry with stimulated echo compensation.&rdquo; *NMR in
 Biomedicine* 29 (4): 431&ndash;43. <https://doi.org/10.1002/nbm.3459>.
 
-Orton, Matthew R., David J. Collins, Dow-Mu Koh, and Martin O. Leach.
-2014. &ldquo;Improved intravoxel incoherent motion analysis of diffusion
+Orton, Matthew R., David J. Collins, Dow-Mu Koh, and Martin O. Leach. 2014. 
+&ldquo;Improved intravoxel incoherent motion analysis of diffusion
 weighted imaging by data driven Bayesian modeling.&rdquo; *Magnetic
 Resonance in Medicine* 71 (1): 411&ndash;20.
 <https://doi.org/10.1002/mrm.24649>.
@@ -441,8 +441,8 @@ and GPU for diagnostic classification of Alzheimer&rsquo;s
 disease.&rdquo; *Frontiers in Neuroinformatics* 7 (January): 50.
 <https://doi.org/10.3389/fninf.2013.00050>.
 
-Stokes, Ashley M., Yesu Feng, Tanya Mitropoulos, and Warren S. Warren.
-2013. &ldquo;Enhanced refocusing of fat signals using optimized
+Stokes, Ashley M., Yesu Feng, Tanya Mitropoulos, and Warren S. Warren. 2013. 
+&ldquo;Enhanced refocusing of fat signals using optimized
 multipulse echo sequences.&rdquo; *Magnetic Resonance in Medicine* 69
 (4): 1044&ndash;55. <https://doi.org/10.1002/mrm.24340>.
 
