### 2019-03-27 mfroeling/QMRITools#11 @mfroeling
i have verified the txt is now readable, im closing the issue

![mac](https://user-images.githubusercontent.com/15890391/55061145-50e29b00-5073-11e9-917b-40f658b417d3.PNG)
