### 2019-04-04 d30708e
update readme
include water component in fat dictionary of EPG fit
