+++ docs/index.md
@@ -48,6 +48,8 @@ language](https://www.wolfram.com/language/) and maintained using
 [eclipse](https://www.eclipse.org/) and runs in the latest version of
 [Wolfram Mathematica](http://www.wolfram.com/mathematica/).
 
+------------------------------------------------------------------------
+
 ## Latest Release
 
 The latesest release can be found
@@ -166,6 +168,8 @@ analyzing data. A summary of the core functionality is listed below.
     -   helix angle and fiber architecture matrix
     -   AHA 17 parameter description
     -   Transmural parameter description
+
+------------------------------------------------------------------------
 
 ## Toolboxes
 
@@ -347,6 +351,12 @@ rejection.](\QMRITools\images\dti.png)
 Import and export functions for tensor data which can be used in the
 [vIST/e](http://bmia.bmt.tue.nl/software/viste/) tractography tool. None
 of the functionality is demonstrated in the `demo.nb`.
+
+------------------------------------------------------------------------
+
+![Full leg diffusion tensor fiber tractography](\QMRITools\images\animation-small.gif)
+
+------------------------------------------------------------------------
 
 ## References
 
