+++ README.md
@@ -53,8 +53,6 @@ All code and documentation is maintained and uploaded to github using [Workbench
 
 ![Guides QMRITools](https://github.com/mfroeling/QMRITools/blob/master/Images/Guide.PNG)
 
-
-
 ## Install toolbox
 
 The latest release can be found [here](https://github.com/mfroeling/QMRITools/releases/download/2.0/QMRITools.zip). 
@@ -88,8 +86,6 @@ To have a global overview of the functionality of the toolbox you can download t
 By default the ``demo.nb`` looks for the folders ``DemoData`` and ``Testing`` in the same folder as the notebook.
 
 In the first section of the demo notebook the toolbox is loaded and two tests are performed. The first test is to check of all files that are needed to run the toolbox are present. The second test runs after the toolbox is loaded and checks if all the functions and their options that are defined are correct.
-
-![Full leg diffusion tensor fiber tractography](https://github.com/mfroeling/QMRITools/blob/master/Images/animation-small.gif)
 
 ## Using the toolbox
 
@@ -191,3 +187,7 @@ program, make sure that you have permission to do so not merely in terms
 of copyright, but also in terms of patent law.
 
 Some code in the NiiTools packages was based on https://github.com/tomdelahaije/nifti-converter
+
+
+![Full leg diffusion tensor fiber tractography](https://github.com/mfroeling/QMRITools/blob/master/Images/animation-small.gif)
+
