### 2019-04-04 #1204 @cMadan
@mfroeling, for now I'd like to wait for the reviewer to follow-up.

@rljacobson, can you check if the last fixes resolve the OS compatibility issue?
