### 2019-04-12 4162fb5
EPG fit update with fat shifting
Included shift calculation in GetPulseProfile
Added ShiftPulseProfile
Added extra methods for ErrorPlot

update Readme with verbose
