+++ QMRITools/Kernel/init.m
@@ -9,17 +9,6 @@
 (*m.froeling@gmail.com*)
 
 
-If[$VersionNumber != 11.3,
-	CreateDialog[Column[{Style["
-	Current Mathematica version is "<>ToString[$VersionNumber]<>"
-	The toolbox is tested developed in 11.3.
-	You need to update! (Or I am behind)
-	Some functions wont work in older versions
-	", TextAlignment -> Center], DefaultButton[], ""}, 
-	Alignment -> Center], WindowTitle -> "Update!"];
-];
-
-
 (*package naem*)
 QMRITools`$Package = "QMRITools`";
 QMRITools`$SubPackages = {
@@ -38,7 +27,7 @@ QMRITools`$SubPackages = {
 
 (*define context and verbose*)
 QMRITools`$Contexts = (QMRITools`$Package <> # & /@ QMRITools`$SubPackages);
-QMRITools`$Verbose = True;
+QMRITools`$Verbose = If[QMRITools`$Verbose===True, True, False];
 
 
 (*print the contexts*)
@@ -90,13 +79,10 @@ If[QMRITools`$Verbose,Print["--------------------------------------"]];
 	
 	SetAttributes[#,{Protected, ReadProtected}]&/@ Names[# <> "*"];
 )& /@ QMRITools`$Contexts;
-(*
-(* ::Section:: *)
-(*Functions*)
 
 
-(*check for latest version*)
-UpdateWarning[]:=If[$VersionNumber != 11.3,
+(*check mathematica version*)
+If[$VersionNumber != 11.3,
 	CreateDialog[Column[{Style["
 	Current Mathematica version is "<>ToString[$VersionNumber]<>"
 	The toolbox is tested developed in 11.3.
@@ -105,111 +91,3 @@ UpdateWarning[]:=If[$VersionNumber != 11.3,
 	", TextAlignment -> Center], DefaultButton[], ""}, 
 	Alignment -> Center], WindowTitle -> "Update!"];
 ];
-
-
-(*Fucntions to clear, load and protect package functions*)
-ClearFunctions[pack_,subpack_,print_:False]:=Module[{packageName,packageSymbols,packageSymbolsG},
-	If[print,Print["--------------------------------------"]];
-	Quiet[
-		If[print,Print["Removing all definitions of "<>#]];
-		packageName = pack <> #;
-		Get[packageName];
-		packageSymbols =Names[packageName <> "*"];
-		packageSymbolsG="Global`"<>#&/@packageSymbols;
-		
-		(*remove all global and private definitions*)
-		Unprotect @@ packageSymbols;
-		ClearAll @@ packageSymbols;
-		Remove @@ packageSymbols;
-		Unprotect @@ packageSymbolsG;
-		ClearAll @@ packageSymbolsG;
-		Remove @@ packageSymbolsG;
-
-		]& /@ subpack;
-];
-
-
-LoadPackages[pack_,subpack_,print_:False]:=Module[{},
-	If[print,Print["--------------------------------------"]];
-	(
-		If[print, Print["Loading all definitions of "<>#]];
-		Get[pack<>#];
-	)&/@subpack;
-]
-
-
-PrintPackages[context_,print_:False]:=Module[{},
-	If[print,
-		Print["--------------------------------------"];
-		Print["All defined packages are: "];
-		Print[context];
-	];
-]
-
-
-ProtectFunctions[pack_,subpack_,print_:False]:=Module[{},
-	If[print,Print["--------------------------------------"]];
-	(
-		If[print,Print["protecting all definitions of "<>#]];
-		SetAttributes[#,{Protected, ReadProtected}]&/@ Names[pack <> # <> "*"];
-		If[print,Print[Names[pack <> # <> "*"]]];
[…]
