+++ README.md
@@ -89,7 +89,8 @@ In the first section of the demo notebook the toolbox is loaded and two tests ar
 
 ## Using the toolbox
 
-The toolbox can be loaded by using `` <<QMRITools` ``
+The toolbox can be loaded by using: `` <<QMRITools` ``
+If you want to monitor the package loading you can use: `` QMRITools`$Verbose = True; <<QMRITools` ``
 
 A list of all QMRITools packages is generated by 
 	
