### 2019-04-29 4a6dc2a
Small imporovementd in functions by using packed array statements
