+++ PacletInfo.m
@@ -1,11 +1,11 @@
 (* Paclet Info File *)
 
-(* created 2019/03/01*)
+(* created 2019/04/12*)
 
 Paclet[
     Name -> "QMRITools",
-    Version -> "2.1.0",
-    MathematicaVersion -> "11.0+",
+    Version -> "2.1.2",
+    MathematicaVersion -> "12.0+",
     Description -> "Toolbox for Quantitative MRI.",
     Creator -> "Martijn Froeling <m.froeling@gmail.com>",
     Thumbnail -> "zthumb.gif",
