### 2019-04-29 9b2f6ad
Small imporovementd in functions by using packed array statements
