### 2019-04-29 #1204 @rljacobson
@mfroeling I have had my own set of technical issues on top of both professional and family distractions. Both of my machines suffered incapacitation for different reasons. @cMadan, I hate to hold up the review of a submission for what is arguably an unreasonable amount of time. If @mfroeling is still willing to wait for my review, I am very happy to do it. On the other hand, he has been superlatively patient with me, and @cMadan might have his own reasons for wanting to move this along.

If I am straining the patience of those involved in this review, one possibility which I would also be agreeable to is for JOSS to accept the submission with the condition that a disclaimer be added that the image registration features are still in beta testing on macOS. Those appear to be the only stumbling block at this stage, and as far as I can tell there is no hard requirement that submitted software work 100% on every platform. If this scenario is agreeable, I would be happy to sign off on the submission and then work with @mfroeling _independently of JOSS_ considering I already am familiar with the issue. On the other hand, again, I am happy to continue my formal review.

My delays are quite unfortunate and unexpected. If I sound excessively apologetic, it is because my delays do not reflect my genuine commitment to the goals and mission of JOSS, nor do they meet my own personal standards of my commitment to my responsibilities as an academic reviewer. But, in this particular case, it was simply unavoidable. I do hope you understand.
