### 2019-05-14 mfroeling/QMRITools#10 @mfroeling
I have validated the mac and unix compatibility.
For future readers if any new issues arise please let me know
