### 2019-05-14 #1204 @mfroeling
@cMadan I have updated the readme regarding the Mac and linux compatibility and het verified that the sripts of the demos run on a Unix server and two different macbooks.
If problems occur with compatibility i am happy to debug them in the future.
For the moment i think i have addressed all issues so i leave it up to you to decide  how to continue.

@rljacobson No problem thanks for the input so far.
