### 2019-05-15 #1204 @cMadan
@mfroeling @rljacobson, thank you both for all your work with these last changes. I will move ahead with acceptance and greatly appreciate @rljacobson's offer to continue to help in working on the OS compatibility issue independent of this JOSS review.

@mfroeling, I will move forward with my final checks and follow-up soon.
