### 2019-05-19 #1204 @cMadan
Still doing checks.

In `README.md`, ``![demo](https://github.com/mfroeling/QMRITools/blob/master/Images/demo.png)`` doesn't match the filename, which is `demo.PNG`.
