+++ PacletInfo.m
@@ -1,10 +1,10 @@
 (* Paclet Info File *)
 
-(* created 2019/04/12*)
+(* created 2019/05/31*)
 
 Paclet[
     Name -> "QMRITools",
-    Version -> "2.1.2",
+    Version -> "2.1.3",
     MathematicaVersion -> "12.0+",
     Description -> "Toolbox for Quantitative MRI.",
     Creator -> "Martijn Froeling <m.froeling@gmail.com>",
