### 2019-06-03 #1204 @cMadan
@mfroeling, everything looks good and I'm almost ready to accept. Can you mint a new release that includes the changes from the review process? I also need a DOI for an archived version of the code (e.g., from Zenodo).
