### 2019-06-03 #1204 @mfroeling
Yes will do i have some small bugs to fix before a new release that i will finalize this week. I don't want to push a new release with known bugs.
