### 2019-06-03 e179b88
bugfix registerdiffusiondatasplit
