### 2019-06-09 #1204 @labarba
Some minor editorial suggestions:

- ¶2: the acronym DWI is used for the first time, without defining (we don't know what that means). Also, recommend to delete "mostly" in first sentence. Add commas after "e.g."
- ¶3: "some demo data which is" >> that is. Comma after "notebook."
- "the documentation is build" >> built
- "such that is incorporated" >> ?
- Cardiac tools: "which allows defining the hard" >> ? (also, two "which allows" in a row)
- "The  first is based on and LMMSE framework" >> delete "and"? should you spell this out?
- Check all other acronyms: can you spell these out on first use?
