### 2019-06-11 af58350
update paper joss
