### 2019-06-11 ecc837b
update paper joss
