### 2019-06-11 #1204 @mfroeling
@labarba The suggested changes are done.
