### 2019-06-11 #1204 @labarba
Big thanks to our editor: @cMadan, and reviewers: @grlee77, @JoshKarpel, @rljacobson — we are grateful for your contribution to JOSS 🙏
